Move chrony config parsing to model::chrony module
Parsing logic is now in the chrony-specific module rather than the
generic Config struct, keeping backend-specific code separate.
