+++ rust/agama-ntp/src/model/chrony.rs
@@ -18,12 +18,15 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
+//! This module defines a chrony-based module for the agama-ntp service.
+
 use agama_software::{Resolvable, ResolvableType};
-use agama_utils::api::ntp::{Config, Source};
+use agama_utils::api::ntp::{Config, Source, SourceType};
 use agama_utils::command::enable_service;
 use async_trait::async_trait;
 use std::path::{Path, PathBuf};
 use std::process::Output;
+use std::str::FromStr;
 use std::{fs, io};
 use tokio::process::Command;
 
@@ -50,22 +53,26 @@ impl Model {
         }
     }
 
+    /// Sets the work directory (usually "/").
     pub fn with_workdir<P: AsRef<Path>>(mut self, workdir: P) -> Self {
         self.workdir = PathBuf::from(workdir.as_ref());
         self
     }
 
+    /// Sets the install directory (usually "/mnt").
     pub fn with_install_dir<P: AsRef<Path>>(mut self, install_dir: P) -> Self {
         self.install_dir = PathBuf::from(install_dir.as_ref());
         self
     }
 
+    /// Chrony configuration path.
     fn config_path(&self) -> PathBuf {
         self.workdir
             .join(CHRONY_CONFIG_DIR)
             .join(CHRONY_CONFIG_FILE)
     }
 
+    /// Chrony configuration path in the installed system.
     fn install_config_path(&self) -> PathBuf {
         self.install_dir
             .join(CHRONY_CONFIG_DIR)
@@ -99,7 +106,7 @@ impl Default for Model {
 impl ModelAdapter for Model {
     async fn get_config(&self) -> Result<Config, Error> {
         let sources_path = self.workdir.join(DRACUT_CHRONY_CONFIG);
-        Ok(Config::from_chrony_conf(sources_path).map_err(Error::ReadConfig)?)
+        parse_chrony_config(sources_path).map_err(Error::ReadConfig)
     }
 
     async fn write_config(&self, config: &Config) -> Result<(), Error> {
@@ -182,6 +189,74 @@ fn generate_chrony_config(sources: &[Source]) -> String {
 
     lines.push(String::new());
     lines.join("\n")
+}
+
+/// Parse NTP configuration from a chrony configuration file.
+///
+/// # Arguments
+///
+/// * `path` - Path to the chrony configuration file
+///
+/// # Returns
+///
+/// A `Config` instance with the parsed sources, or an error if reading/parsing fails.
+pub fn parse_chrony_config<P: AsRef<Path>>(path: P) -> io::Result<Config> {
+    let content = fs::read_to_string(path)?;
+    parse_chrony_config_str(&content)
+}
+
+/// Parse NTP configuration from a chrony configuration string.
+///
+/// # Arguments
+///
+/// * `content` - The chrony configuration content
+///
+/// # Returns
+///
+/// A `Config` instance with the parsed sources, or an error if parsing fails.
+pub fn parse_chrony_config_str(content: &str) -> io::Result<Config> {
+    let mut sources = Vec::new();
+
+    for line in content.lines() {
+        let line = line.trim();
+
+        // Skip empty lines and comments
+        if line.is_empty() || line.starts_with('#') {
+            continue;
+        }
+
+        let parts: Vec<&str> = line.split_whitespace().collect();
+        if parts.len() < 2 {
+            continue;
+        }
+
+        let source_type = match SourceType::from_str(parts[0]) {
+            Ok(st) => st,
+            Err(_) => continue, // Skip unknown source types
+        };
+
+        let address = parts[1].to_string();
+        let mut iburst = false;
+        let mut offline = false;
+
+        // Parse options
+        for option in &parts[2..] {
+            match *option {
+                "iburst" => iburst = true,
+                "offline" => offline = true,
+                _ => {} // Ignore unknown options
+            }
+        }
+
+        sources.push(Source {
+            source_type,
+            address,
+            iburst,
+            offline,
+        });
+    }
+
+    Ok(Config { sources })
 }
 
 // Errors from chronyc are logged to stdout. This ancillary function turns an Output object into a
@@ -343,5 +418,123 @@ mod tests {
             .await
             .expect("Failed to remove the chrony configuration");
         assert!(!written_path.exists());
+    }
+
+    #[test]
+    fn test_parse_chrony_config_str_simple() {
+        let content = "pool 0.opensuse.pool.ntp.org iburst\n";
+        let config = parse_chrony_config_str(content).unwrap();
+
+        assert_eq!(config.sources.len(), 1);
+        assert_eq!(config.sources[0].source_type, SourceType::Pool);
+        assert_eq!(config.sources[0].address, "0.opensuse.pool.ntp.org");
+        assert_eq!(config.sources[0].iburst, true);
+        assert_eq!(config.sources[0].offline, false);
+    }
+
+    #[test]
+    fn test_parse_chrony_config_str_multiple_sources() {
+        let content = r#"# Generated by Agama
+pool 0.opensuse.pool.ntp.org iburst
+server ntp.example.com offline
+peer ntp-peer.local
+"#;
+        let config = parse_chrony_config_str(content).unwrap();
+
+        assert_eq!(config.sources.len(), 3);
+
+        assert_eq!(config.sources[0].source_type, SourceType::Pool);
+        assert_eq!(config.sources[0].address, "0.opensuse.pool.ntp.org");
+        assert_eq!(config.sources[0].iburst, true);
+        assert_eq!(config.sources[0].offline, false);
+
+        assert_eq!(config.sources[1].source_type, SourceType::Server);
+        assert_eq!(config.sources[1].address, "ntp.example.com");
+        assert_eq!(config.sources[1].iburst, false);
+        assert_eq!(config.sources[1].offline, true);
+
+        assert_eq!(config.sources[2].source_type, SourceType::Peer);
+        assert_eq!(config.sources[2].address, "ntp-peer.local");
+        assert_eq!(config.sources[2].iburst, false);
+        assert_eq!(config.sources[2].offline, false);
+    }
+
+    #[test]
+    fn test_parse_chrony_config_str_with_both_options() {
+        let content = "server ntp.example.com iburst offline\n";
+        let config = parse_chrony_config_str(content).unwrap();
+
+        assert_eq!(config.sources.len(), 1);
+        assert_eq!(config.sources[0].iburst, true);
+        assert_eq!(config.sources[0].offline, true);
+    }
+
+    #[test]
+    fn test_parse_chrony_config_str_ignores_comments_and_empty_lines() {
+        let content = r#"
+# This is a comment
+pool ntp1.org
+
+# Another comment
+server ntp2.org iburst
+
+"#;
+        let config = parse_chrony_config_str(content).unwrap();
+
+        assert_eq!(config.sources.len(), 2);
+        assert_eq!(config.sources[0].address, "ntp1.org");
+        assert_eq!(config.sources[1].address, "ntp2.org");
+    }
+
+    #[test]
+    fn test_parse_chrony_config_str_empty_content() {
+        let content = "# Only comments\n\n";
+        let config = parse_chrony_config_str(content).unwrap();
+
+        assert_eq!(config.sources.len(), 0);
+    }
+
+    #[test]
+    fn test_parse_chrony_config_str_invalid_lines_skipped() {
+        let content = r#"pool ntp.org iburst
+invalid_line
+server ntp2.org
+just_one_word
+"#;
+        let config = parse_chrony_config_str(content).unwrap();
+
+        assert_eq!(config.sources.len(), 2);
+        assert_eq!(config.sources[0].address, "ntp.org");
+        assert_eq!(config.sources[1].address, "ntp2.org");
+    }
+
+    #[test]
+    fn test_parse_chrony_config_str_unknown_options_ignored() {
+        let content = "pool ntp.org iburst unknown_option offline another_unknown\n";
+        let config = parse_chrony_config_str(content).unwrap();
+
+        assert_eq!(config.sources.len(), 1);
+        assert_eq!(config.sources[0].iburst, true);
+        assert_eq!(config.sources[0].offline, true);
+    }
+
+    #[test]
+    fn test_parse_chrony_config_file() {
+        use std::io::Write;
+        use tempfile::NamedTempFile;
+
+        let content = r#"# Test config
+pool 0.opensuse.pool.ntp.org iburst
+server ntp.example.com offline
+"#;
+
+        let mut temp_file = NamedTempFile::new().unwrap();
+        temp_file.write_all(content.as_bytes()).unwrap();
+
+        let config = parse_chrony_config(temp_file.path()).unwrap();
+
+        assert_eq!(config.sources.len(), 2);
+        assert_eq!(config.sources[0].source_type, SourceType::Pool);
+        assert_eq!(config.sources[1].source_type, SourceType::Server);
     }
 }
