+++ rust/agama-ntp/src/service.rs
@@ -89,7 +89,7 @@ impl Service {
     /// Initializes the service by reading the current configuration.
     pub async fn setup(&mut self) {
         if let Ok(config) = self.model.get_config().await {
-            if !config.is_empty() {
+            if !config.sources.is_empty() {
                 self.config = Some(config);
             }
         }
