+++ rust/agama-utils/src/api/ntp/config.rs
@@ -21,9 +21,6 @@
 use merge::Merge;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::fs;
-use std::io;
-use std::path::Path;
 use std::str::FromStr;
 
 /// NTP configuration.
@@ -35,81 +32,6 @@ pub struct Config {
     #[serde(default)]
     #[merge(strategy = overwrite_non_empty)]
     pub sources: Vec<Source>,
-}
-
-impl Config {
-    /// Read NTP configuration from a chrony configuration file.
-    ///
-    /// # Arguments
-    ///
-    /// * `path` - Path to the chrony configuration file
-    ///
-    /// # Returns
-    ///
-    /// A `Config` instance with the parsed sources, or an error if reading/parsing fails.
-    pub fn from_chrony_conf<P: AsRef<Path>>(path: P) -> io::Result<Self> {
-        let content = fs::read_to_string(path)?;
-        Self::from_chrony_conf_str(&content)
-    }
-
-    /// Parse NTP configuration from a chrony configuration string.
-    ///
-    /// # Arguments
-    ///
-    /// * `content` - The chrony configuration content
-    ///
-    /// # Returns
-    ///
-    /// A `Config` instance with the parsed sources, or an error if parsing fails.
-    pub fn from_chrony_conf_str(content: &str) -> io::Result<Self> {
-        let mut sources = Vec::new();
-
-        for line in content.lines() {
-            let line = line.trim();
-
-            // Skip empty lines and comments
-            if line.is_empty() || line.starts_with('#') {
-                continue;
-            }
-
-            let parts: Vec<&str> = line.split_whitespace().collect();
-            if parts.len() < 2 {
-                continue;
-            }
-
-            let source_type = match SourceType::from_str(parts[0]) {
-                Ok(st) => st,
-                Err(_) => continue, // Skip unknown source types
-            };
-
-            let address = parts[1].to_string();
-            let mut iburst = false;
-            let mut offline = false;
-
-            // Parse options
-            for option in &parts[2..] {
-                match *option {
-                    "iburst" => iburst = true,
-                    "offline" => offline = true,
-                    _ => {} // Ignore unknown options
-                }
-            }
-
-            sources.push(Source {
-                source_type,
-                address,
-                iburst,
-                offline,
-            });
-        }
-
-        Ok(Config { sources })
-    }
-
-    /// Determines whether the configuration is empty.
-    pub fn is_empty(&self) -> bool {
-        self.sources.is_empty()
-    }
 }
 
 fn overwrite_non_empty<T>(left: &mut Vec<T>, mut right: Vec<T>) {
@@ -258,104 +180,6 @@ mod tests {
     }
 
     #[test]
-    fn test_from_chrony_conf_str_simple() {
-        let content = "pool 0.opensuse.pool.ntp.org iburst\n";
-        let config = Config::from_chrony_conf_str(content).unwrap();
-
-        assert_eq!(config.sources.len(), 1);
-        assert_eq!(config.sources[0].source_type, SourceType::Pool);
-        assert_eq!(config.sources[0].address, "0.opensuse.pool.ntp.org");
-        assert_eq!(config.sources[0].iburst, true);
-        assert_eq!(config.sources[0].offline, false);
-    }
-
-    #[test]
-    fn test_from_chrony_conf_str_multiple_sources() {
-        let content = r#"# Generated by Agama
-pool 0.opensuse.pool.ntp.org iburst
-server ntp.example.com offline
-peer ntp-peer.local
-"#;
-        let config = Config::from_chrony_conf_str(content).unwrap();
-
-        assert_eq!(config.sources.len(), 3);
-
-        assert_eq!(config.sources[0].source_type, SourceType::Pool);
-        assert_eq!(config.sources[0].address, "0.opensuse.pool.ntp.org");
-        assert_eq!(config.sources[0].iburst, true);
-        assert_eq!(config.sources[0].offline, false);
-
-        assert_eq!(config.sources[1].source_type, SourceType::Server);
-        assert_eq!(config.sources[1].address, "ntp.example.com");
-        assert_eq!(config.sources[1].iburst, false);
-        assert_eq!(config.sources[1].offline, true);
-
-        assert_eq!(config.sources[2].source_type, SourceType::Peer);
-        assert_eq!(config.sources[2].address, "ntp-peer.local");
-        assert_eq!(config.sources[2].iburst, false);
-        assert_eq!(config.sources[2].offline, false);
-    }
-
-    #[test]
-    fn test_from_chrony_conf_str_with_both_options() {
-        let content = "server ntp.example.com iburst offline\n";
-        let config = Config::from_chrony_conf_str(content).unwrap();
-
-        assert_eq!(config.sources.len(), 1);
-        assert_eq!(config.sources[0].iburst, true);
-        assert_eq!(config.sources[0].offline, true);
-    }
-
-    #[test]
-    fn test_from_chrony_conf_str_ignores_comments_and_empty_lines() {
-        let content = r#"
-# This is a comment
-pool ntp1.org
-
-# Another comment
-server ntp2.org iburst
-
-"#;
-        let config = Config::from_chrony_conf_str(content).unwrap();
-
-        assert_eq!(config.sources.len(), 2);
-        assert_eq!(config.sources[0].address, "ntp1.org");
-        assert_eq!(config.sources[1].address, "ntp2.org");
-    }
-
-    #[test]
-    fn test_from_chrony_conf_str_empty_content() {
-        let content = "# Only comments\n\n";
-        let config = Config::from_chrony_conf_str(content).unwrap();
-
-        assert_eq!(config.sources.len(), 0);
-    }
-
-    #[test]
-    fn test_from_chrony_conf_str_invalid_lines_skipped() {
-        let content = r#"pool ntp.org iburst
-invalid_line
-server ntp2.org
-just_one_word
-"#;
-        let config = Config::from_chrony_conf_str(content).unwrap();
-
-        assert_eq!(config.sources.len(), 2);
-        assert_eq!(config.sources[0].address, "ntp.org");
-        assert_eq!(config.sources[1].address, "ntp2.org");
-    }
-
-    #[test]
-    fn test_from_chrony_conf_str_unknown_options_ignored() {
-        let content = "pool ntp.org iburst unknown_option offline another_unknown\n";
-        let config = Config::from_chrony_conf_str(content).unwrap();
-
-        assert_eq!(config.sources.len(), 1);
-        assert_eq!(config.sources[0].iburst, true);
-        assert_eq!(config.sources[0].offline, true);
-    }
-
-    #[test]
     fn test_source_type_from_str() {
         assert_eq!(SourceType::from_str("pool").unwrap(), SourceType::Pool);
         assert_eq!(SourceType::from_str("server").unwrap(), SourceType::Server);
@@ -363,25 +187,5 @@ just_one_word
         assert_eq!(SourceType::from_str("Pool").unwrap(), SourceType::Pool);
         assert_eq!(SourceType::from_str("SERVER").unwrap(), SourceType::Server);
         assert!(SourceType::from_str("unknown").is_err());
-    }
-
-    #[test]
-    fn test_from_chrony_conf_file() {
-        use std::io::Write;
-        use tempfile::NamedTempFile;
-
-        let content = r#"# Test config
-pool 0.opensuse.pool.ntp.org iburst
-server ntp.example.com offline
-"#;
-
-        let mut temp_file = NamedTempFile::new().unwrap();
-        temp_file.write_all(content.as_bytes()).unwrap();
-
-        let config = Config::from_chrony_conf(temp_file.path()).unwrap();
-
-        assert_eq!(config.sources.len(), 2);
-        assert_eq!(config.sources[0].source_type, SourceType::Pool);
-        assert_eq!(config.sources[1].source_type, SourceType::Server);
     }
 }
